+++ dotnet-6/Services/UserService.cs
@@ -2,14 +2,6 @@ using CrudApp.DataAccess;
 using CrudApp.Models;
 
 namespace CrudApp.Services;
-public interface IUserService 
-{
-    IEnumerable<User> GetUsers();
-    User GetById(long id);
-    User CreateOrUpdate(User user);
-    void Delete(long id);
-
-}
 
 public class UserService : IUserService
 {
